Added tf support for auto-conversion to ss
Now if `f` is a `TransferFunction`, and `g` is a `StateSpace`, all
operations will automatically promote the arguments to `StateSpace`, and
then perform the computation. This allows `f + g` to work, for example.

Also added a crude implementation of `pole` for `TransferFunction`, and
made all the plotting and time-response calculations work for them as
well.