Added frequency response functionality
New functions:

`freqresp`, `evalfr`: evaluate the frequency response

`bode`, `nyquist`, `sigma`: compute specific aspects of the frequency
response. `*plot` versions of the above are also defined.

`zpkdata`: extract the zeros, poles, and gain of the system.

`gain`: extrac the gain of the system

`markovparam`: compute the markov parameters of a state-space system

In the process of writing these functions, a few bugs were found in old
functionality. Most importantly: an error in `Poly` with the setting of
`nzfirst` was found for `Poly(0)`. A small error in time-series plotting
was also fixed.