Update README.md
The type inference for the array of transfer functions with a comprehension does not seem to be reliable. There is a note about this in the Julia documention http://docs.julialang.org/en/release-0.4/manual/arrays/
For me, without controlling the type explicitly, CLs is inferred as Array{Any,1}, which stepplot does not accept.