Explicit type parameters in return val of c2d
The `c2d` function is complex enough that inference often struggles with the return type (not for Float64, but for more complex types, e.g., `Dual(Unitful)` etc.). I believe it should be safe to assume that the numeric type of the returned object is the `T` that was determined as the promotion of all the matrix element types, with automatic conversion as fallback.