Use foreground color to plot unit circle in pzmap
By changing the :seriestype from :path to :shape the shape's stroke
color is defined by :foreground_color_subplot instead of :linecolor
(shapes are rendered similar to plot markers). This way setting
:linecolor explicitely can be omitted and the stroke color is defined by
the plot theme or can be even adjusted by the user. This is useful, when
using dark themes for instance.