remove automatic field propagation from `balance_statespace`
The reason is that many abstract statespace types have fields that are affected by the balancing, and the automatic forwarding of those fields without application of the balancing led to incorrect answers. The proper way of handling this is for each type to implement their own method of `balance_statespace` if the additional fields are required.